agents/manager/requirements.txt
pymongo==4.5.0
dnspython==2.4.2
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.4
requests==2.31.0
fastapi==0.100.0
uvicorn==0.23.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
yfinance==0.2.28
jinja2==3.1.2
pydantic==2.4.2
anthropic==0.19.1
matplotlib==3.7.2
seaborn==0.12.2
httpx==0.24.1